+++ NEWS.md
@@ -2,6 +2,8 @@
 
 - Load test data from *MsDataHub* dropping dependency from the deprecated
   *msdata* package.
+- Reduce code duplication in `gplot,XChromatograms`/`gplot,XChromatogram`
+  functions.
 
 ## Changes in v0.99.5 (commit: b419691)